pyproject.toml
[tool.poetry]
name = "fujida_agent"
version = "0.1.0"
description = ""
authors = ["Elisey <eliseyblajen@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
pydantic-settings = "^2.10.1"
aiogram = "^3.21.0"
pgvector = "^0.4.1"
asyncpg = "^0.30.0"
alembic = "^1.16.4"
sqlalchemy = "^2.0.41"
python-dotenv = "^1.1.1"
phonenumbers = "^9.0.10"
openai = "^1.97.1"
httpx = "^0.28.1"
ffmpeg = "^1.4"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
